chore(javascript): use full precision of "float" type values
Values of type float are internally represented with double precision,
so don't downcast to float unnecessarily.

+++ src/net/sourceforge/kolmafia/textui/javascript/ValueConverter.java
@@ -200,7 +200,7 @@ public class ValueConverter {
         || object instanceof Double d && JavascriptNumbers.isDoubleSafeInteger(d)) {
       return DataTypes.makeIntValue(((Number) object).longValue());
     } else if (object instanceof Float || object instanceof Double) {
-      return DataTypes.makeFloatValue(((Number) object).floatValue());
+      return DataTypes.makeFloatValue(((Number) object).doubleValue());
     } else if (object instanceof String) {
       return DataTypes.makeStringValue((String) object);
     } else if (object instanceof StringBuffer || object instanceof ConsString) {
